Enforce some code standards when compiling C.
- Use C99 standard.
- Enable feature test macro for POSIX.1-2008.
- Treat warnings as errors.

Fix all code that does not conform to these standards.

+++ login/crypto.c
@@ -52,7 +52,7 @@ static int login_tag(bool verify, const char* host_id, const char* action,
     free(message);
     return -1;
   }
-  if (ret >= message_len) {
+  if ((size_t) ret >= message_len) {
     free(message);
     return -1;
   }
